Decompose Go reference support (#4109)
