+++ changelog.d/unreleased/4109.internal.md
@@ -1,0 +1,16 @@
+---
+category: internal
+issues:
+  - 4109
+affected:
+  - src/CodeIndex/Indexer/References/Support/LanguageReferenceExtractionSupport.cs
+  - src/CodeIndex/Indexer/References/Support/LanguageReferenceExtractionSupport.Go.cs
+---
+
+## English
+
+- **Go reference support now lives in a focused partial file (#4109)** - moved the Go import-block mapping, branch-label references, and type-reference helpers out of the shared language support file without changing reference extraction behavior.
+
+## 日本語
+
+- **Go 参照サポートを専用 partial ファイルに分離しました (#4109)** - Go の import block マップ、branch label 参照、型参照ヘルパーを共有言語サポート本体から移動し、参照抽出の挙動は変更していません。
